backend/requirements.txt
fastapi
pydantic
uvicorn
python-jose
passlib
# spark-nlp==5.4.2
# pyspark==3.4.2
# spark-nlp
# pyspark
numpy
nltk
bert-extractive-summarizer
torch